requirements.txt
cmake==3.27.9
langchain==0.1.10
langchain-community==0.0.25
langchain-core==0.1.28
langchain-text-splitters==0.0.1
pypdf==4.0.2
PyPDF2==3.0.1
python-pptx==0.6.23
python-docx==1.1.0
llama-cpp-python==0.1.78
numpy==1.24.1
pinecone-client==2.2.4